backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.0.350
openai==1.3.5
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4 